+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Release history
 
+#### 2.0.0
+Complete refactor. Instead of creating default listeners add an ability to create custom listeners and
+target groups. The loadbalancer now exposes an explicit method to easily create a pair of listeners and
+a pair of target groups for easy out-of-the-box blue/green deployment configuration.
+
 #### 1.0.2
 Renames.